\alias{TENxBrainData}
\name{TENxBrainData}

\title{10X Brain Data}

\description{Single-cell RNA-seq data for 1.3 million brain cells from E18 mice, generated by 10X Genomics.}

\usage{
TENxBrainData()
}

\value{
A SingleCellExperiment object with a HDF5Matrix in the \code{counts} assay, which contains UMI counts for each gene in each cell.
Row- and column-level metadata are also preserved.
}

\details{
Single-cell RNA-seq data were generated by 10X Genomics using the Cromium Megacell technology, and processed into counts using CellRanger 1.3.0.
Cells were obtained from the cortex, hippocampus and subventricular zone of two E18 mice.
}

\seealso{
\code{\link[SingleCellExperiment]{SingleCellExperiment}}
}

\author{
Aaron Lun
}

\references{
10X Genomics (2017).
1.3 Million Brain Cells from E18 Mice.
\url{https://support.10xgenomics.com/single-cell-gene-expression/datasets/1.3.0/1M_neurons}
}
