\alias{TENxBrainData}
\name{TENxBrainData}

\title{10X Brain Data}

\description{Single-cell RNA-seq data for 1.3 million brain cells from E18 mice, generated by 10X Genomics.}

\usage{
TENxBrainData()
}

\value{
A SingleCellExperiment object with a HDF5Matrix in the \code{counts} assay, which contains UMI counts for each gene in each cell.
Row- and column-level metadata are also provided.
}

\details{
Single-cell RNA-seq data were generated by 10X Genomics using the Cromium Megacell technology and processed using CellRanger 1.3.0.
Cells were obtained from the cortex, hippocampus and subventricular zone of two E18 mice (1306127 cells in total).
Each count represents the number of unique molecular identifiers (UMIs) assigned to each gene in the Ensembl annotation (27998 genes in total).

The Ensembl ID and gene symbol are provided in the row-level metadata.
The barcode sequence, sequencing library ID and mouse of origin are provided in the column-level metadata.
}

\seealso{
\code{\link[SingleCellExperiment]{SingleCellExperiment}}
}

\author{
Aaron Lun
}

\references{
10X Genomics (2017).
1.3 Million Brain Cells from E18 Mice.
\url{https://support.10xgenomics.com/single-cell-gene-expression/datasets/1.3.0/1M_neurons}
}

\examples{
\dontrun{
sce <- TENxBrainData()
sce
}
}
